refactor: split aces_processor/compiler.py into a package (#779)
* refactor: split aces_processor/compiler.py into a package

Convert the 2677-line compiler module into an API-stable compiler/ package of
<=500-line submodules (foundations -> domain compilers -> pipeline -> facade).
No behavior change: public API (compile_runtime_model,
compile_scenario_runtime_model) and every external import line are unchanged,
and the existing test suite passes untouched. Drop the oversized-allowlist
entry and remove three unreachable private helpers. Update doc/spec path
references to the deleted file. Part of #3.

* Fix SonarCloud new-code findings from compiler split

The file split re-scans the moved code as new code, so pre-existing
maintainability issues surfaced as new violations. Resolve them without
behavior change: replace bare `Any` parameter hints with the concrete SDL
types (`object` for the polymorphic `_dump`), collapse the two outcome-address
if-chains into dict lookups, fold `_resolve_node_ref`'s require_vm/require_switch
into a single `required_type` (fewer params and returns), bundle
`_compile_workflow_step`'s loop-invariant arguments into a context object, and
decompose `_compile_action_contracts` into small extraction helpers. The test
suite is unchanged and green.
