+++ implementations/python/packages/aces_processor/compiler/__init__.py
@@ -1,0 +1,11 @@
+"""SDL-to-runtime compiler.
+
+Public facade for the compiler package. The compilation logic is split across
+package-private submodules (foundations -> domain compilers -> pipeline); this
+module re-exports the stable public API so ``aces_processor.compiler`` keeps its
+external contract.
+"""
+
+from .pipeline import compile_runtime_model, compile_scenario_runtime_model
+
+__all__ = ["compile_runtime_model", "compile_scenario_runtime_model"]
