+++ specs/formal/objectives/README.md
@@ -46,7 +46,7 @@ target-resolution, success-interpretation, and dependency-ordering semantics
   (`_verify_objectives`)
 - compiled runtime objective resource, addresses, diagnostics, and
   ordering/refresh derivation:
-  - `implementations/python/packages/aces_processor/compiler.py`
+  - `implementations/python/packages/aces_processor/compiler/`
   - `implementations/python/packages/aces_processor/models/`
 - planner ordering/refresh reconciliation:
   - `implementations/python/packages/aces_processor/planner.py`
