+++ specs/formal/workflows/compensation.md
@@ -99,7 +99,7 @@ workflow can be `cancelled` or `timed_out` while compensation is `running` or
 
 - SDL authoring models: `implementations/python/packages/aces_sdl/orchestration.py`
 - semantic validation: `implementations/python/packages/aces_sdl/validator/`
-- compiled contracts: `implementations/python/packages/aces_processor/compiler.py`
+- compiled contracts: `implementations/python/packages/aces_processor/compiler/`
 - runtime state/history models: `implementations/python/packages/aces_processor/models/`
 - manager validation: `implementations/python/packages/aces_processor/manager.py`
 - control-plane lifecycle handling: `implementations/python/packages/aces_processor/control_plane.py`
